Improve running codelens commands in a split
Previously, running an LSP code lens command would open a new split and
terminal emulator each time a command was run. This change allows us to
reuse an open window and reuse a scratch buffer.
It turns out that doing this kind of window managment requires a fair
bit of intervention, so we introduce a `util` module to abstract some of
the more complicated or frequent bits.
One potential problem with this is that running concurrent code lens
command would cause their outputs to overwrite each other.
We can consider making this use plenary.job in the future

Remove vestigial check for valid buffer when reporting job failure